49: setup yarn, get rid of globals
