+++ env.example.php
@@ -4,9 +4,15 @@
  * - create a file named env.php
  * - copy the content of this file
  * - change the variables according to your environement */
-
-$db_host = "localhost"; // SQL Database host
-$db_name = "database_name"; // SQL Database name
-$db_user = "root"; // SQL Database username
-$db_password = ""; // SQL Database password
-$debug_mode = false; // Keep this false for production
+return [
+    // SQL Database host:
+    "db_host" => "localhost",
+    // SQL Database name:
+    "db_name" => "mydatabase",
+    // SQL Database username:
+    "db_user" => "root",
+    // SQL Database password:
+    "db_password" => "",
+    // Keep this false for production:
+    "developement" => true
+];
